api testing/python/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
flask
sqlalchemy